modify end_points paths

+++ app/controllers/items_controller.rb
@@ -52,17 +52,18 @@ class ItemsController < ApplicationController
 
     # Only allow a trusted parameter "white list" through.
     def item_params
-      params.require(:item).permit(:title, :price, :stock, :brand, :description, :images => [])
+      params.require(:item).permit(:title, :price, :stock, :brand, :description, :images => [], :category_id, :sub_categories => [])
     end
 
     # Only allow a trusted parameter "white list" through.
     def cat_params
-      params.permit(:category_id, :sub_categories => [])
+      params.require(:item).permit()
     end
 
     def item_category
-      @category = Category.find_by(id: cat_params[:category_id])
-      cat_params[:sub_categories].each do |sub_category|
+      return false if item_params[:category_id].nil? || item_params[:sub_categories].nil?
+      @category = Category.find_by(id: item_params[:category_id])
+      item_params[:sub_categories].each do |sub_category|
         @subcat = SubCategory.new(title: sub_category, category_id: @category.id, item_id: @item.id)
         unless @subcat.save
           SubCategory.where(item_id: @item.id).destroy_all
